refactor(api): extract fund admin endpoints into dedicated FundsController
Keeps GivingController's primary constructor unchanged so existing tests
compile without edits. Routes unchanged (api/v1/giving/admin/funds/*).
