+++ src/Koinon.Api/Controllers/GivingController.cs
@@ -19,7 +19,6 @@ namespace Koinon.Api.Controllers;
 public class GivingController(
     IBatchDonationEntryService batchDonationService,
     IContributionStatementService statementService,
-    IFundService fundService,
     ILogger<GivingController> logger) : ControllerBase
 {
     #region Funds
@@ -72,137 +71,6 @@ public class GivingController(
         logger.LogInformation("Retrieved fund: {IdKey}", idKey);
 
         return Ok(new { data = result.Value });
-    }
-
-    /// <summary>
-    /// Gets all funds for administration (active and inactive).
-    /// </summary>
-    /// <param name="ct">Cancellation token</param>
-    /// <returns>List of all funds with full admin detail</returns>
-    /// <response code="200">Returns list of all funds</response>
-    [HttpGet("admin/funds")]
-    [RequiresClaim("financial", "view")]
-    [ProducesResponseType(typeof(IReadOnlyList<FundAdminDto>), StatusCodes.Status200OK)]
-    public async Task<IActionResult> GetAllFundsAdminAsync(CancellationToken ct = default)
-    {
-        var funds = await fundService.GetAllFundsAsync(ct);
-
-        logger.LogInformation("Retrieved {Count} funds for admin", funds.Count);
-
-        return Ok(new { data = funds });
-    }
-
-    /// <summary>
-    /// Creates a new fund.
-    /// </summary>
-    /// <param name="request">The fund creation request</param>
-    /// <param name="ct">Cancellation token</param>
-    /// <returns>The created fund</returns>
-    /// <response code="201">Fund created successfully</response>
-    /// <response code="400">Invalid request</response>
-    [HttpPost("admin/funds")]
-    [RequiresClaim("financial", "edit")]
-    [ProducesResponseType(typeof(FundAdminDto), StatusCodes.Status201Created)]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    public async Task<IActionResult> CreateFundAsync(
-        [FromBody] CreateFundRequest request,
-        CancellationToken ct = default)
-    {
-        var result = await fundService.CreateFundAsync(request, ct);
-
-        if (!result.IsSuccess)
-        {
-            logger.LogWarning("Failed to create fund: {Error}", result.Error);
-            return Problem(
-                detail: result.Error?.Message ?? "Failed to create fund",
-                statusCode: StatusCodes.Status400BadRequest,
-                title: "Bad Request"
-            );
-        }
-
-        logger.LogInformation("Created fund: {IdKey}", result.Value!.IdKey);
-
-        return StatusCode(StatusCodes.Status201Created, new { data = result.Value });
-    }
-
-    /// <summary>
-    /// Updates an existing fund.
-    /// </summary>
-    /// <param name="idKey">The fund IdKey</param>
-    /// <param name="request">The fund update request</param>
-    /// <param name="ct">Cancellation token</param>
-    /// <returns>The updated fund</returns>
-    /// <response code="200">Fund updated successfully</response>
-    /// <response code="400">Invalid request</response>
-    /// <response code="404">Fund not found</response>
-    [HttpPut("admin/funds/{idKey}")]
-    [ValidateIdKey]
-    [RequiresClaim("financial", "edit")]
-    [ProducesResponseType(typeof(FundAdminDto), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> UpdateFundAsync(
-        string idKey,
-        [FromBody] UpdateFundRequest request,
-        CancellationToken ct = default)
-    {
-        var result = await fundService.UpdateFundAsync(idKey, request, ct);
-
-        if (!result.IsSuccess)
-        {
-            logger.LogWarning("Failed to update fund {IdKey}: {Error}", idKey, result.Error);
-
-            if (result.Error?.Code == "NOT_FOUND")
-            {
-                return Problem(
-                    detail: result.Error?.Message ?? "Fund not found",
-                    statusCode: StatusCodes.Status404NotFound,
-                    title: "Not Found"
-                );
-            }
-
-            return Problem(
-                detail: result.Error?.Message ?? "Failed to update fund",
-                statusCode: StatusCodes.Status400BadRequest,
-                title: "Bad Request"
-            );
-        }
-
-        logger.LogInformation("Updated fund: {IdKey}", idKey);
-
-        return Ok(new { data = result.Value });
-    }
-
-    /// <summary>
-    /// Deactivates a fund (soft delete — sets IsActive = false).
-    /// </summary>
-    /// <param name="idKey">The fund IdKey</param>
-    /// <param name="ct">Cancellation token</param>
-    /// <returns>No content on success</returns>
-    /// <response code="204">Fund deactivated successfully</response>
-    /// <response code="404">Fund not found</response>
-    [HttpDelete("admin/funds/{idKey}")]
-    [ValidateIdKey]
-    [RequiresClaim("financial", "edit")]
-    [ProducesResponseType(StatusCodes.Status204NoContent)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> DeactivateFundAsync(string idKey, CancellationToken ct = default)
-    {
-        var result = await fundService.DeactivateFundAsync(idKey, ct);
-
-        if (!result.IsSuccess)
-        {
-            logger.LogWarning("Failed to deactivate fund {IdKey}: {Error}", idKey, result.Error);
-            return Problem(
-                detail: result.Error?.Message ?? "Fund not found",
-                statusCode: StatusCodes.Status404NotFound,
-                title: "Not Found"
-            );
-        }
-
-        logger.LogInformation("Deactivated fund: {IdKey}", idKey);
-
-        return NoContent();
     }
 
     #endregion
